--- components/heroes/shaders/simulationVertexShader.js
@@ -1,16 +1,0 @@
-const vertexShader = `
-varying vec2 vUv;
-
-void main() {
-  vUv = uv;
-
-  vec4 modelPosition = modelMatrix * vec4(position, 1.0);
-  vec4 viewPosition = viewMatrix * modelPosition;
-  vec4 projectedPosition = projectionMatrix * viewPosition;
-
-  gl_Position = projectedPosition;
-}
-
-`
-
-export default vertexShader
